backend/pyproject.toml
[project]
name = "retrynote-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.27.1,<1.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.27,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "alembic>=1.13.1,<2.0.0",
    # Auth
    "PyJWT>=2.8.0,<3.0.0",
    "bcrypt>=4.1.2,<5.0.0",
    # Validation
    "pydantic>=2.6.1,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "email-validator>=2.1.0,<3.0.0",
    # Queue
    "celery[redis]>=5.3.6,<6.0.0",
    "redis>=5.0.1,<6.0.0",
    # AI
    "openai>=1.13.0,<2.0.0",
    "google-genai>=1.0.0,<2.0.0",
    # File processing
    "PyPDF2>=3.0.1,<4.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "python-pptx>=0.6.23,<1.0.0",
    "Pillow>=10.2.0,<11.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "httpx>=0.27.0,<1.0.0",
    "markdown>=3.5.2,<4.0.0",
    # Search
    "pgvector>=0.2.4,<1.0.0",
    # Storage
    "boto3>=1.34.0,<2.0.0",
    # Security
    "slowapi>=0.1.9,<1.0.0",
    "python-magic>=0.4.27,<1.0.0",
    "disposable-email-domains>=0.0.1",
    # Utils
    "python-dateutil>=2.8.2,<3.0.0",
    "pytz>=2024.1",
    "stripe>=10.0.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.2,<9.0.0",
    "pytest-asyncio>=0.23.5,<1.0.0",
]


[dependency-groups]
dev = [
    "pytest>=8.0.2,<9.0.0",
    "pytest-asyncio>=0.23.5,<1.0.0",
    "pytest-xdist>=3.8.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "ruff>=0.15.9",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
